ocr-service/requirements.txt
# Python OCR microservice dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9   # Required for FastAPI file uploads
pytesseract==0.3.13
Pillow==10.3.0
pdf2image==1.17.0
